[MCJ-1355] refactor: 도메인형 아키텍처로 패키지 구조 재편
- domain/groupbuy/{domain,application,presentation} 구조로 재배치
- domain/store/{infrastructure,application,presentation} 구조로 재배치
- global/BaseEntity(ours) 제거 → global/entity/BaseEntity(팀원) 통일
- BaseEntity nullable 필드 반영: createdAt/updatedAt → LocalDateTime?
- GroupBuyRequestService: changedAt 바인딩 시 null-safe 처리 추가
