+++ src/main/kotlin/com/moongchijang/domain/groupbuy/domain/entity/GroupBuyRequestStatus.kt
@@ -1,0 +1,5 @@
+package com.moongchijang.domain.groupbuy.domain.entity
+
+enum class GroupBuyRequestStatus {
+    SUBMITTED, REVIEWING, CONTACTING, OPENED, REJECTED
+}
