+++ src/main/kotlin/com/moongchijang/domain/groupbuy/domain/repository/GroupBuyRequestRepository.kt
@@ -1,6 +1,6 @@
-package com.moongchijang.domain.groupbuy.repository
+package com.moongchijang.domain.groupbuy.domain.repository
 
-import com.moongchijang.domain.groupbuy.entity.GroupBuyRequest
+import com.moongchijang.domain.groupbuy.domain.entity.GroupBuyRequest
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface GroupBuyRequestRepository : JpaRepository<GroupBuyRequest, Long> {
