--- src/main/kotlin/com/moongchijang/domain/groupbuy/entity/GroupBuyRequestStatus.kt
@@ -1,9 +1,0 @@
-package com.moongchijang.domain.groupbuy.entity
-
-enum class GroupBuyRequestStatus {
-    SUBMITTED,
-    REVIEWING,
-    CONTACTING,
-    OPENED,
-    REJECTED
-}
